feat(support): add AppiumImage class
this class wraps a Jimp and means we do not need to monkeypatch a Jimp instance. it has consistent behavior around promises.  I did not simply _proxy_ `Jimp` due to the complexity involved.

this deprecates `getJimpImage`.  It's possible there are some things that drivers need out of the `Jimp` instance that `AppiumImage` doesn't provide, so we will need to look closely and add wrappers.